pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Michał Wiśniewski"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2.12"
psycopg2-binary = "^2.9.3"
djangorestframework = "3.13.1"
django-allauth = "^0.49.0"
dj-rest-auth = "^2.2.3"
celery = "^5.2.3"
redis = "^4.2.0"
django-celery-beat = "^2.2.1"
watchdog = "^2.1.6"
PyYAML = "^6.0"
drf-yasg = "^1.20.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.0"
pytest-django = "^4.5.2"
pytest-cov = "^3.0.0"
black = "^22.1.0"
isort = "^5.10.1"
pytest-faker = "^2.0.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
skip-string-normalization = true
extend-exclude =  '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''
